refactor: extract handlers from SchemaService (886 -> 324 lines)
Extract 9 standalone handler classes from SchemaService:
- SchemaHandler, CommentHandler, ExtensionHandler
- EnumHandler, SequenceHandler, FunctionHandler
- ProcedureHandler, ViewHandler, TriggerHandler

Each handler encapsulates statement generation for one object type.
SchemaService now acts as an orchestration facade.

Closes #46
