+++ src/core/schema/handlers/index.ts
@@ -1,0 +1,9 @@
+export { CommentHandler } from "./comment-handler";
+export { EnumHandler } from "./enum-handler";
+export { ExtensionHandler } from "./extension-handler";
+export { FunctionHandler } from "./function-handler";
+export { ProcedureHandler } from "./procedure-handler";
+export { SchemaHandler } from "./schema-handler";
+export { SequenceHandler } from "./sequence-handler";
+export { TriggerHandler } from "./trigger-handler";
+export { ViewHandler } from "./view-handler";
